Move appsetings to utils
