+++ apps/vs-code-designer/src/app/utils/appSettings/encryptLocalSettings.ts
