+++ apps/vs-code-designer/src/app/utils/appSettings/getLocalSettingsFile.ts
