+++ apps/vs-code-designer/src/app/utils/appSettings/localSettings.ts
@@ -2,10 +2,10 @@
  *  Copyright (c) Microsoft Corporation. All rights reserved.
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
-import { localSettingsFileName } from '../../constants';
-import { localize } from '../../localize';
-import { writeFormattedJson } from '../utils/fs';
-import { parseJson } from '../utils/parseJson';
+import { localSettingsFileName } from '../../../constants';
+import { localize } from '../../../localize';
+import { writeFormattedJson } from '../fs';
+import { parseJson } from '../parseJson';
 import { DialogResponses, parseError } from '@microsoft/vscode-azext-utils';
 import type { IActionContext } from '@microsoft/vscode-azext-utils';
 import type { ILocalSettingsJson } from '@microsoft/vscode-extension';
